Move ANSI terminal color support into loggocolor
As discussed on the juju-dev mailing list, it is problematic
that the loggo package emits ANSI escape codes by default.

This PR moves the color support into a sub-package that
can be used optionally when required.
